+++ src/types/user.ts
@@ -2,5 +2,5 @@ export interface IUser {
     id: string;
     username: string;
     age: number;
-    hobbies: string[];
+    hobbies?: string[];
 }
